requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

aioredis==1.3.1
asgiref==3.4.1 ; python_version >= '3.6'
async-timeout==4.0.1 ; python_version >= '3.6'
attrs==21.2.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
autobahn==21.3.1 ; python_version >= '3.7'
automat==20.2.0
backports.entry-points-selectable==1.1.1 ; python_version >= '2.7'
boto3==1.20.3
botocore==1.23.3 ; python_version >= '3.6'
certifi==2021.10.8
cffi==1.15.0
channels-redis==3.3.1
channels==3.0.4
charset-normalizer==2.0.7 ; python_version >= '3'
constantly==15.1.0
cryptography==35.0.0 ; python_version >= '3.6'
daphne==3.0.2 ; python_version >= '3.6'
distlib==0.3.3
django-countries==7.2.1
django-dotenv==1.4.2
django-seed==0.3.1
django-storages==1.12.3
django==4.0b1
djangorestframework==3.12.4
faker==9.8.0 ; python_version >= '3.6'
filelock==3.3.2 ; python_version >= '3.6'
hiredis==2.0.0 ; python_version >= '3.6'
hyperlink==21.0.0
idna==3.3 ; python_version >= '3'
incremental==21.3.0
jmespath==0.10.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
msgpack==1.0.2
pathspec==0.9.0
pbr==5.7.0 ; python_version >= '2.6'
pillow==8.4.0
pip==21.3.1 ; python_version >= '3.6'
pipenv-to-requirements==0.9.0
pipenv==2021.11.9 ; python_version >= '3.6'
platformdirs==2.4.0 ; python_version >= '3.6'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pyopenssl==21.0.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0 ; python_version >= '3.6'
sentry-sdk==1.4.3
service-identity==21.1.0
setuptools==58.5.3 ; python_version >= '3.6'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2 ; python_version >= '3.5'
text-unidecode==1.3
toposort==1.7
twisted-iocpsupport==1.0.2 ; platform_system == 'Windows'
twisted[tls]==21.7.0 ; python_full_version >= '3.6.7'
txaio==21.2.1 ; python_version >= '3.6'
typing-extensions==3.10.0.2
tzdata==2021.5 ; sys_platform == 'win32'
urllib3==1.26.7 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
virtualenv-clone==0.5.7 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
virtualenv==20.10.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zope.interface==5.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
psycopg2==2.9.1